+++ packages/integration-tests/tests/recommendations.test.js
@@ -1,7 +1,7 @@
 /*
  * © 2021 Thoughtworks, Inc.
  */
-import { Selector, ClientFunction } from 'testcafe'
+import { ClientFunction } from 'testcafe'
 import waitOn from 'wait-on'
 import page from './page-model'
 const getLocation = ClientFunction(() => document.location.href)
@@ -31,9 +31,7 @@ test('filter components render with correct data when app loads', async (t) => {
   await t.expect(page.recommendationTypes.exists).ok()
 })
 
-test('card components render with correct data when app loads', async (t) => {
-  // commented out because the error message is not always present
-  // await t.expect(page.errorMessage.exists).ok()
+test('card components render with data when app loads', async (t) => {
   await t.expect(page.lastThirtyDayTotal.exists).ok()
   await t.expect(page.projectedThirtyDayTotal.exists).ok()
   await t.expect(page.forecastEquivalencyCard.exists).ok()
@@ -41,34 +39,20 @@ test('card components render with correct data when app loads', async (t) => {
   await t.expect(page.costSavingsPerMonth.exists).ok()
 })
 
-test('table components render with correct data when app loads', async (t) => {
+test('table components renders with data when app loads', async (t) => {
   await t.expect(page.searchInput.exists).ok()
   await t.expect(page.recommendationsDataGrid.exists).ok()
 })
 
 test('toggle changes unit of measure', async (t) => {
-  //check projected totals
-  // commented out because the error message is not always present
-  // await t.expect(page.errorMessage.exists).ok()
-
-  await t
-    .expect(page.unitOfMeasureLastThirtyDayTotal.textContent)
-    .eql('Metric Tons CO2e')
-  await t
-    .expect(page.unitOfMeasureProjectedThirtyDayTotal.textContent)
-    .eql('Metric Tons CO2e')
-  await t.expect(page.co2eSavingsLastThirtyDayTotal.exists).ok()
-  await t.expect(page.co2eSavingsProjectedThirtyDayTotal.exists).ok()
-  await t.expect(page.costSavingsLastThirtyDayTotal.exists).ok()
-  await t.expect(page.costSavingsProjectedThirtyDayTotal.exists).ok()
   //check first cell
   await t
     .expect(page.tableSavingsColumn.textContent)
     .eql('Potential Carbon Savings (t)')
   await t.expect(page.firstSavingsCell.exists).ok()
   //click kilogram toggle
-  await t.click(page.toggle, { isTrusted: true })
-  //recheck data in - kg instead of metric tons, so 1000
+  await t.click(page.toggle)
+  // recheck data in - kg instead of metric tons, so 1000
   await t
     .expect(page.unitOfMeasureLastThirtyDayTotal.textContent)
     .eql('Kilograms CO2e')