+++ packages/api/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @cloud-carbon-footprint/api
 
+## 1.8.0
+
+### Minor Changes
+
+- 6ffe7497: adds electricity maps api handling. For CLI create-app template changes, refer to [this commit](https://github.com/cloud-carbon-footprint/cloud-carbon-footprint/commit/2749723c00343865c7453fea29609ea6dad4b5e9).
+
+### Patch Changes
+
+- Updated dependencies [bc06b861]
+- Updated dependencies [7630768d]
+- Updated dependencies [dd98c8cb]
+- Updated dependencies [6ffe7497]
+  - @cloud-carbon-footprint/common@1.13.0
+  - @cloud-carbon-footprint/app@1.4.0
+
+## 1.7.0
+
+### Minor Changes
+
+- fedf79c0: API and CLI processes now persist connection to mongodb client when enabled as cache
+
+  For Create-App changes, please refer to this [commit](https://github.com/cloud-carbon-footprint/cloud-carbon-footprint/commit/8995b8a7f29fb06f8a437166d32e75b1ed147870).
+
+- 045ef419: Adds OpenAPI spec documentation and Swagger portal
+
+### Patch Changes
+
+- Updated dependencies [1a5d7636]
+- Updated dependencies [c82bf5fd]
+- Updated dependencies [fedf79c0]
+  - @cloud-carbon-footprint/common@1.12.0
+  - @cloud-carbon-footprint/app@1.3.0
+
 ## 1.6.2
 
 ### Patch Changes